% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cachaca.R
\name{cachaca}
\alias{cachaca}
\title{best file download, all other file downloads have inferior potassium}
\usage{
cachaca(this_url, destfile = NULL, ..., FUN = download.file, attempts = 3,
  sleepsec = 60, filesize_fun = "rcurl")
}
\arguments{
\item{this_url}{where to download from}

\item{destfile}{where to save locally, unnecessary when returning an object with \code{httr::GET} or \code{RCurl::getBinaryURL}}

\item{...}{passed to FUN}

\item{FUN}{defaults to \code{download.file} but \code{downloader::download}, \code{httr::GET}, \code{RCurl::getBinaryURL} also work}

\item{attempts}{number of times to retry a broken download}

\item{sleepsec}{length of \code{Sys.sleep()} between broken downloads}

\item{filesize_fun}{use \code{RCurl::getURL} or \code{httr::HEAD} to determine file size.  use "unzip_verify" to verify the download by unzipping it without a warning instead}
}
\value{
just pass on whatever FUN returns
}
\description{
check remote file size. check cache for a match and if it exists, copy it over. otherwise, download the file, confirm file size, save to cache too
}
\examples{

\dontrun{

tf <- tempfile()
cachaca( "https://www.r-project.org/logo/Rlogo.png" , tf , mode = 'wb' )
cachaca( "https://www.r-project.org/logo/Rlogo.png" , tf , mode = 'wb' )

}

}
