% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_catalog.R
\name{get_catalog}
\alias{get_catalog}
\title{retrieve a listing of all available extracts for a microdata set}
\usage{
get_catalog(data_name, output_dir = getwd(), ...)
}
\arguments{
\item{data_name}{a character vector with a microdata abbreviation}

\item{output_dir}{home directory on the local computer to save the microdata}

\item{...}{passed to \code{get_catalog}}
}
\value{
\code{data.frame} with files and folders available for download
}
\description{
subheader
}
\examples{

\dontrun{

# download the entire american time use survey catalog
my_catalog <- get_catalog( "atus" )

# review files from data year 2012
subset( my_catalog , directory == 2012 )

setwd( "C:/My Directory/ATUS" )

# download data year 2012
lodown( "atus" , subset( my_catalog , directory == 2012 ) )

list.files()

}

}
\author{
Anthony Damico
}

