% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_catalog.R, R/lodown.R
\name{get_catalog}
\alias{get_catalog}
\alias{lodown}
\title{locally download, import, prepare publicly-available microdata for analysis}
\usage{
get_catalog(data_name, output_dir = getwd(), ...)

lodown(data_name, catalog = NULL, ...)
}
\arguments{
\item{data_name}{a character vector with a microdata abbreviation}

\item{output_dir}{home directory on the local computer to save the microdata}

\item{...}{passed to \code{get_catalog} and \code{lodown_}}

\item{catalog}{\code{data.frame} detailing available microdata extracts}
}
\value{
TRUE if it worked
}
\description{
get_catalog retrieves a listing of all available extracts for a microdata set
}
\details{
stores microdata in the current working directory
}
\examples{

\dontrun{

# examples to download everything
lodown( "ahrf" , output_dir = "C:/My Directory/AHRF" )
lodown( "atus" , output_dir = "C:/My Directory/ATUS" )
lodown( "ess" , output_dir = "C:/My Directory/ESS" , your_email = "email@address.com" )
lodown( "nis" , output_dir = "C:/My Directory/NIS" )

# examples to download only the first two records in the catalog
ahrf_cat <- get_catalog( "ahrf" , output_dir = "C:/My Directory/AHRF" )
lodown( "ahrf" , ahrf_cat[ 1:2 , ] )
atus_cat <- get_catalog( "atus" , output_dir = "C:/My Directory/ATUS" )
lodown( "atus" , atus_cat[ 1:2 , ] )
ess_cat <- get_catalog( "ess" , output_dir = "C:/My Directory/ESS" )
lodown( "ess" , ess_cat[ 1:2 , ] , your_email = "email@address.com" )
nis_cat <- get_catalog( "nis" , output_dir = "C:/My Directory/NIS" )
lodown( "nis" , nis_cat[ 1:2 , ] )

}

}
\author{
Anthony Damico
}

