% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icpsr.R
\name{lodown_icpsr}
\alias{lodown_icpsr}
\alias{get_catalog_icpsr}
\title{locally download, import, prepare icpsr microdata}
\usage{
lodown_icpsr(data_name, catalog, ...)

get_catalog_icpsr(series_number = NULL, study_numbers = NULL,
  archive = "ICPSR", bundle_preference = c("rdata", "stata", "sas", "spss",
  "delimited", "ascsas", "ascstata", "ascspss", "ascii", "gis", "doc"))
}
\arguments{
\item{data_name}{a character vector with a microdata abbreviation}

\item{catalog}{\code{data.frame} detailing available microdata extracts}

\item{...}{passed to \code{get_catalog} and \code{lodown_}}

\item{series_number}{any data series number from http://www.icpsr.umich.edu/icpsrweb/ICPSR/series}

\item{study_numbers}{one or more study numbers from http://www.icpsr.umich.edu/icpsrweb/ICPSR/studies?q=}

\item{archive}{defaults to "ICPSR" but other archives include "HMCA" , "NACJD" , "NAHDAP" , "NADAC" , "civicleads" , "RCMD" , "ADDEP" , "childcare" , "NCAA" , "NACDA" , "DSDR" , "METLDB"}

\item{bundle_preference}{when multiple filetypes are available, which should be given priority?}
}
\description{
get_catalog_icpsr retrieves a listing of all available extracts for a microdata set
}
