% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_functions.R
\name{MIsvychisq}
\alias{MIsvychisq}
\title{variant of survey::svychisq that works on multiply-imputed data}
\usage{
MIsvychisq(formula, design, statistic = "Chisq", ...)
}
\description{
variant of survey::svychisq that works on multiply-imputed data
}
\seealso{
\url{http://r-survey.r-forge.r-project.org/survey/html/svychisq.html}
}
