% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_functions.R
\name{MIsvyciprop}
\alias{MIsvyciprop}
\title{variant of survey::svyciprop that works on multiply-imputed data}
\usage{
MIsvyciprop(formula, design, method = c("logit", "likelihood", "asin", "beta",
  "mean", "xlogit"), level = 0.95, df = mean(unlist(lapply(design$designs,
  survey::degf))), ...)
}
\description{
variant of survey::svyciprop that works on multiply-imputed data
}
\seealso{
\url{http://r-survey.r-forge.r-project.org/survey/html/svyciprop.html}
}
