% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_functions.R
\name{MIsvyttest}
\alias{MIsvyttest}
\title{variant of survey::svyttest that works on multiply-imputed data}
\usage{
MIsvyttest(formula, design, ...)
}
\description{
variant of survey::svyttest that works on multiply-imputed data
}
\seealso{
\url{http://r-survey.r-forge.r-project.org/survey/html/svyttest.html}
}
