% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nls.R
\name{nls_get_selections}
\alias{nls_get_selections}
\alias{nls_get_weights}
\title{functions to access the national longitudinal studies panel weights}
\usage{
nls_get_selections(study)

nls_get_weights(study, uona, selections)
}
\arguments{
\item{study}{must be one of the options shown at \url{https://www.nlsinfo.org/weights} like \code{c( "nlsy97" , "nlsy79" , "nlscya" , "nlsym" , "nlsom" , "nlsyw" , "nlsmw" )}}

\item{uona}{use "NO" if you want to weight using "the respondents are in ALL of the selected years" or "YES" for "the respondents are in ANY OR ALL of the selected years"}

\item{selections}{panel timepoints to use, as returned by \code{nls_get_selections}}
}
\description{
download all available survey-year selections for any of the nlsy studies, then download the appropriate set of weights
}
\seealso{
\url{http://www.asdfree.com} for usage examples
}
