% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scf.R
\name{pirls_MIcombine}
\alias{pirls_MIcombine}
\title{multiply-imputed plausible value calculations for the Progress in International Reading Literacy Study}
\usage{
pirls_MIcombine(results, variances, call = sys.call(), df.complete = Inf,
  ...)
}
\description{
multiply-imputed plausible value calculations for the Progress in International Reading Literacy Study
}
\seealso{
\url{https://cran.r-project.org/web/packages/mitools/mitools.pdf}
}
