% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbo.R
\name{sbo_MIcombine}
\alias{sbo_MIcombine}
\alias{sbo_with}
\alias{sbo_subset}
\alias{sbo_update}
\alias{sbo_degf}
\alias{sbo_MIsvyciprop}
\title{dual design calculations for the survey of business owners}
\usage{
sbo_MIcombine(x, adjustment = 1.992065)

sbo_with(sbo.svy, expr, ...)

sbo_subset(x, ...)

sbo_update(x, ...)

sbo_degf(x)

sbo_MIsvyciprop(formula, design, method = c("logit", "likelihood", "asin",
  "beta", "mean", "xlogit"), level = 0.95,
  df = mean(unlist(lapply(design$designs, survey::degf))), ...)
}
\description{
the \code{mitools::MIcombine} variant includes a 2007-specific variance adjustment.  this will change in other years.
this adjustment statistic was pulled from the middle of page 8
\url{https://www2.census.gov/econ/sbo/07/pums/2007_sbo_pums_users_guide.pdf#page=8}
}
\details{
each of these sbo-specific functions contain a variant of some other \code{library(survey)} function that also maintains the census bureau's dual design calculation.
these functions expect both the coefficient and the variance survey objects
}
\seealso{
\url{https://cran.r-project.org/web/packages/mitools/mitools.pdf}
}
