% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scf.R
\name{scf_MIcombine}
\alias{scf_MIcombine}
\title{variant of \code{mitools::MIcombine} that only uses the sampling variance from the first implicate instead of averaging all five}
\usage{
scf_MIcombine(results, variances, call = sys.call(), df.complete = Inf, ...)
}
\description{
variant of \code{mitools::MIcombine} that only uses the sampling variance from the first implicate instead of averaging all five
}
\seealso{
\url{https://cran.r-project.org/web/packages/mitools/mitools.pdf}
}
