% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scf.R
\name{timss_MIcombine}
\alias{timss_MIcombine}
\title{multiply-imputed plausible value calculations for the Trends in International Mathematics and Science Study}
\usage{
timss_MIcombine(results, variances, call = sys.call(), df.complete = Inf,
  ...)
}
\description{
multiply-imputed plausible value calculations for the Trends in International Mathematics and Science Study
}
\seealso{
\url{https://cran.r-project.org/web/packages/mitools/mitools.pdf}
}
