% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cachaca.R
\name{cachaca}
\alias{cachaca}
\title{best file download, all other file downloads have inferior potassium}
\usage{
cachaca(this_url, destfile = NULL, ..., FUN = download.file, attempts = 3,
  sleepsec = 60, filesize_fun = "httr",
  savecache = getOption("lodown.cachaca.savecache", TRUE),
  cdc_ftp_https = TRUE)
}
\arguments{
\item{this_url}{where to download from}

\item{destfile}{where to save locally, unnecessary when returning an object with \code{httr::GET} or \code{RCurl::getBinaryURL}}

\item{...}{passed to FUN}

\item{FUN}{defaults to \code{download.file} but \code{downloader::download}, \code{httr::GET}, \code{RCurl::getBinaryURL} also work}

\item{attempts}{number of times to retry a broken download}

\item{sleepsec}{length of \code{Sys.sleep()} between broken downloads}

\item{filesize_fun}{use \code{httr::HEAD} or \code{RCurl::getURL} to determine file size.  use "unzip_verify" to verify the download by attempting to unzip the file without issue}

\item{savecache}{whether to actually cache the downloaded files in the temporary directories.  setting this option to FALSE eliminates the purpose of cachaca(), but sometimes it's necessary to disable for a single call, or globally.}

\item{cdc_ftp_https}{whether to substitute cdc ftp sites with https, i.e. \code{gsub( "ftp://ftp.cdc.gov/" , "https://ftp.cdc.gov/" , this_url , fixed = TRUE )}.}
}
\value{
just pass on whatever FUN returns
}
\description{
check remote file size. check cache for a match and if it exists, copy it over. otherwise, download the file, confirm file size, save to cache too
}
\examples{

\dontrun{

# only the first of these two lines downloads the file.  the second line uses the cached file
tf <- tempfile()
cachaca( "https://www.r-project.org/logo/Rlogo.png" , tf , mode = 'wb' )
cachaca( "https://www.r-project.org/logo/Rlogo.png" , tf , mode = 'wb' )

# option to disable cache in case the files are too big
options( "lodown.cachaca.savecache" = FALSE )
cachaca( "https://www.r-project.org/logo/Rlogo.svg" , tf , mode = 'wb' )
cachaca( "https://www.r-project.org/logo/Rlogo.svg" , tf , mode = 'wb' )

}

}
