% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lodown.R
\name{lodown}
\alias{lodown}
\title{locally download, import, prepare publicly-available microdata for analysis}
\usage{
lodown(data_name, catalog = NULL, ...)
}
\arguments{
\item{data_name}{a character vector with a microdata abbreviation}

\item{catalog}{a \code{data.frame} containing folders and files to load}

\item{...}{passed to \code{get_catalog} and \code{lodown_}}
}
\value{
TRUE if it worked
}
\description{
subheader
}
\details{
stores microdata in the current working directory
}
\examples{

\dontrun{

setwd( "C:/My Directory/ATUS" )

lodown( "atus" )

}

}
\author{
Anthony Damico
}

