% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lodown.R
\name{lodown}
\alias{get_catalog}
\alias{lodown}
\title{locally download, import, prepare publicly-available microdata}
\usage{
lodown(data_name, catalog = NULL, ...)

get_catalog(data_name, output_dir = getwd(), ...)
}
\arguments{
\item{data_name}{a character vector with a microdata abbreviation}

\item{catalog}{\code{data.frame} detailing available microdata extracts}

\item{...}{passed to \code{get_catalog} and \code{lodown_}}

\item{output_dir}{directory on your local computer to save the microdata}
}
\value{
a freshly-prepared microdata extract on your local computer
}
\description{
get_catalog retrieves a listing of all available extracts for a microdata set
}
\examples{

\dontrun{

# examples to download everything
lodown( "ahrf" , output_dir = "C:/My Directory/AHRF" )
lodown( "atus" , output_dir = "C:/My Directory/ATUS" )
lodown( "ess" , output_dir = "C:/My Directory/ESS" , your_email = "email@address.com" )
lodown( "nis" , output_dir = "C:/My Directory/NIS" )
lodown( "nsch" , output_dir = "C:/My Directory/NSCH" )
lodown( "yrbss" , output_dir = "C:/My Directory/YRBSS" )

# examples to download only the first two records in the catalog
ahrf_cat <- get_catalog( "ahrf" , output_dir = "C:/My Directory/AHRF" )
lodown( "ahrf" , ahrf_cat[ 1:2 , ] )
atus_cat <- get_catalog( "atus" , output_dir = "C:/My Directory/ATUS" )
lodown( "atus" , atus_cat[ 1:2 , ] )
ess_cat <- get_catalog( "ess" , output_dir = "C:/My Directory/ESS" )
lodown( "ess" , ess_cat[ 1:2 , ] , your_email = "email@address.com" )
nis_cat <- get_catalog( "nis" , output_dir = "C:/My Directory/NIS" )
lodown( "nis" , nis_cat[ 1:2 , ] )
nsch_cat <- get_catalog( "nsch" , output_dir = "C:/My Directory/NSCH" )
lodown( "nsch" , nsch_cat[ 1:2 , ] )
yrbss_cat <- get_catalog( "yrbss" , output_dir = "C:/My Directory/YRBSS" )
lodown( "yrbss" , yrbss_cat[ 1:2 , ] )
#'
}

}
\author{
Anthony Damico
}

