% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lodown.R
\name{lodown}
\alias{get_catalog}
\alias{lodown}
\title{locally download, import, prepare publicly-available microdata}
\usage{
lodown(data_name, catalog = NULL, ...)

get_catalog(data_name, output_dir = getwd(), ...)
}
\arguments{
\item{data_name}{a character vector with a microdata abbreviation}

\item{catalog}{\code{data.frame} detailing available microdata extracts}

\item{...}{passed to \code{get_catalog} and \code{lodown_}}

\item{output_dir}{directory on your local computer to save the microdata}
}
\value{
TRUE, and also the microdata in either the folder you specified or your working directory
}
\description{
get_catalog retrieves a listing of all available extracts for a microdata set
}
\examples{

\dontrun{

# examples to download everything
lodown( "acs" , output_dir = "C:/My Directory/ACS" )
lodown( "addhealth" , output_dir = "C:/My Directory/AddHealth" , 
		your_email = "email@address.com" , your_password = "password" )
lodown( "ahrf" , output_dir = "C:/My Directory/AHRF" )
lodown( "anes" , output_dir = "C:/My Directory/ANES" , your_email = "email@address.com" )
lodown( "atus" , output_dir = "C:/My Directory/ATUS" )
lodown( "brfss" , output_dir = "C:/My Directory/BRFSS" )
lodown( "ces" , output_dir = "C:/My Directory/CES" )
lodown( "ess" , output_dir = "C:/My Directory/ESS" , your_email = "email@address.com" )
lodown( "faers" , output_dir = "C:/My Directory/FAERS" )
lodown( "ncvs" , output_dir = "C:/My Directory/NCVS" ,
		your_email = "email@address.com" , your_password = "password" )
lodown( "nhis" , output_dir = "C:/My Directory/NHIS" )
lodown( "nis" , output_dir = "C:/My Directory/NIS" )
lodown( "nibrs" , output_dir = "C:/My Directory/NIBRS" ,
		your_email = "email@address.com" , your_password = "password" )
lodown( "nsch" , output_dir = "C:/My Directory/NSCH" )
lodown( "nsfg" , output_dir = "C:/My Directory/NSFG" )
lodown( "nppes" , output_dir = "C:/My Directory/NPPES" )
lodown( "nsduh" , output_dir = "C:/My Directory/NSDUH" , 
		your_email = "email@address.com" , your_password = "password" )
lodown( "pme" , output_dir = "C:/My Directory/PME" )
lodown( "pnadc" , output_dir = "C:/My Directory/PNADC" )
lodown( "pof" , output_dir = "C:/My Directory/POF" )
lodown( "sbo" , output_dir = "C:/My Directory/SBO" )
lodown( "scf" , output_dir = "C:/My Directory/SCF" )
lodown( "yrbss" , output_dir = "C:/My Directory/YRBSS" )

# examples to download only the first two records in the catalog
acs <- get_catalog( "acs" , output_dir = "C:/My Directory/ACS" )
lodown( "acs" , acs[ 1:2 , ] )
addhealth_cat <- get_catalog( "addhealth" , output_dir = "C:/My Directory/AddHealth" )
lodown( "addhealth" , addhealth_cat , 
		your_email = "email@address.com" , your_password = "password" )
ahrf_cat <- get_catalog( "ahrf" , output_dir = "C:/My Directory/AHRF" )
lodown( "ahrf" , ahrf_cat[ 1:2 , ] )
anes_cat <- get_catalog( "anes" , output_dir = "C:/My Directory/ANES" , 
		your_email = "email@address.com" )
lodown( "anes" , anes_cat , your_email = "email@address.com" )
atus_cat <- get_catalog( "atus" , output_dir = "C:/My Directory/ATUS" )
lodown( "atus" , atus_cat[ 1:2 , ] )
brfss_cat <- get_catalog( "brfss" , output_dir = "C:/My Directory/BRFSS" )
lodown( "brfss" , brfss_cat[ 1:2 , ] )
ces_cat <- get_catalog( "ces" , output_dir = "C:/My Directory/CES" )
lodown( "ces" , ces_cat[ 1:2 , ] )
ess_cat <- get_catalog( "ess" , output_dir = "C:/My Directory/ESS" )
lodown( "ess" , ess_cat[ 1:2 , ] , your_email = "email@address.com" )
faers_cat <- get_catalog( "faers" , output_dir = "C:/My Directory/FAERS" )
lodown( "faers" , faers_cat[ 1:2 , ] )
ncvs_cat <- get_catalog( "ncvs" , output_dir = "C:/My Directory/NCVS" )
lodown( "ncvs" , ncvs_cat[ 1:2 , ] ,
		your_email = "email@address.com" , your_password = "password" )
nhis_cat <- get_catalog( "nhis" , output_dir = "C:/My Directory/NHIS" )
lodown( "nhis" , nhis_cat[ 1:2 , ] )
nibrs_cat <- get_catalog( "nibrs" , output_dir = "C:/My Directory/NIBRS" )
lodown( "nibrs" , nibrs_cat[ 1:2 , ] ,
		your_email = "email@address.com" , your_password = "password" )
nis_cat <- get_catalog( "nis" , output_dir = "C:/My Directory/NIS" )
lodown( "nis" , nis_cat[ 1:2 , ] )
nppes_cat <- get_catalog( "nppes" , output_dir = "C:/My Directory/NPPES" )
lodown( "nppes" , nppes_cat )
nsch_cat <- get_catalog( "nsch" , output_dir = "C:/My Directory/NSCH" )
lodown( "nsch" , nsch_cat[ 1:2 , ] )
nsfg_cat <- get_catalog( "nsfg" , output_dir = "C:/My Directory/NSFG" )
lodown( "nsfg" , nsfg_cat[ 1:2 , ] )
nsduh_cat <- get_catalog( "nsduh" , output_dir = "C:/My Directory/NSDUH" )
lodown( "nsduh" , nsduh_cat[ 1:2 , ] ,
		your_email = "email@address.com" , your_password = "password" )
pme_cat <- get_catalog( "pme" , output_dir = "C:/My Directory/PME" )
lodown( "pme" , pme_cat[ 1:2 , ] )
pnadc_cat <- get_catalog( "pnadc" , output_dir = "C:/My Directory/PNADC" )
lodown( "pnadc" , pnadc_cat[ 1:2 , ] )
pof_cat <- get_catalog( "pof" , output_dir = "C:/My Directory/POF" )
lodown( "pof" , pof_cat[ 1:2 , ] )
sbo_cat <- get_catalog( "sbo" , output_dir = "C:/My Directory/SBO" )
lodown( "sbo" , sbo_cat[ 1:2 , ] )
scf_cat <- get_catalog( "scf" , output_dir = "C:/My Directory/SCF" )
lodown( "scf" , scf_cat[ 1:2 , ] )
yrbss_cat <- get_catalog( "yrbss" , output_dir = "C:/My Directory/YRBSS" )
lodown( "yrbss" , yrbss_cat[ 1:2 , ] )

}

}

