% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lodown.R
\name{get_catalog}
\alias{get_catalog}
\alias{lodown}
\title{locally download and prepare publicly-available microdata}
\usage{
get_catalog(data_name, output_dir = getwd(), ...)

lodown(data_name, catalog = NULL, ...)
}
\arguments{
\item{data_name}{a character vector with a microdata abbreviation}

\item{output_dir}{directory on your local computer to save the microdata}

\item{...}{passed to \code{get_catalog_*} and \code{lodown_*}}

\item{catalog}{\code{data.frame} detailing available microdata extracts and metadata}
}
\value{
catalog \code{data.frame} detailing available microdata extracts and metadata, along with local file paths and possibly additional metadata acquired during the download
}
\description{
\code{lodown} actively downloads the extracts specified by the catalog.  \code{get_catalog} retrieves a listing of all available extracts for a microdata set.  skipping \code{get_catalog} will attempt to download all available extracts
}
\seealso{
\url{http://www.asdfree.com} for usage examples
}
